refactor(orc8r): Segregation of Builder service as internal
