+++ cwf/cloud/go/services/cwf/test_init/test_service_init.go
@@ -17,14 +17,14 @@ import (
 	"testing"
 
 	"magma/cwf/cloud/go/services/cwf"
-	"magma/cwf/cloud/go/services/cwf/servicers"
+	builder_servicers "magma/cwf/cloud/go/services/cwf/servicers/protected"
 	"magma/orc8r/cloud/go/orc8r"
 	builder_protos "magma/orc8r/cloud/go/services/configurator/mconfig/protos"
 	"magma/orc8r/cloud/go/test_utils"
 )
 
 func StartTestService(t *testing.T) {
-	StartTestServiceInternal(t, servicers.NewBuilderServicer())
+	StartTestServiceInternal(t, builder_servicers.NewBuilderServicer())
 }
 
 func StartTestServiceInternal(t *testing.T, builder builder_protos.MconfigBuilderServer) {
