+++ lte/cloud/go/services/lte/test_init/test_service_init.go
@@ -55,7 +55,7 @@ func StartTestServiceWithConfig(t *testing.T, serviceConfig lte_service.Config) 
 	}
 
 	srv, lis := test_utils.NewTestOrchestratorService(t, lte.ModuleName, lte_service.ServiceName, labels, annotations)
-	builder_protos.RegisterMconfigBuilderServer(srv.GrpcServer, servicers.NewBuilderServicer(serviceConfig))
+	builder_protos.RegisterMconfigBuilderServer(srv.GrpcServer, builder_servicers.NewBuilderServicer(serviceConfig))
 	provider_protos.RegisterStreamProviderServer(srv.GrpcServer, servicers.NewProviderServicer())
 
 	// Init storage
