+++ orc8r/cloud/go/services/orchestrator/test_init/test_service_init.go
@@ -20,6 +20,7 @@ import (
 	builder_protos "magma/orc8r/cloud/go/services/configurator/mconfig/protos"
 	"magma/orc8r/cloud/go/services/orchestrator"
 	"magma/orc8r/cloud/go/services/orchestrator/servicers"
+	builder_servicers "magma/orc8r/cloud/go/services/orchestrator/servicers/protected"
 	indexer_protos "magma/orc8r/cloud/go/services/state/protos"
 	streamer_protos "magma/orc8r/cloud/go/services/streamer/protos"
 	streamer_servicers "magma/orc8r/cloud/go/services/streamer/servicers"
@@ -37,7 +38,7 @@ func (srv *testStreamerServer) GetUpdates(req *protos.StreamRequest, stream prot
 }
 
 func StartTestService(t *testing.T) {
-	StartTestServiceInternal(t, servicers.NewBuilderServicer(), servicers.NewIndexerServicer(), servicers.NewProviderServicer())
+	StartTestServiceInternal(t, builder_servicers.NewBuilderServicer(), servicers.NewIndexerServicer(), servicers.NewProviderServicer())
 }
 
 func StartTestServiceInternal(
